+++ slack-invite/api/endpoints/slack_endpoint.rb
@@ -88,10 +88,10 @@ module Api
                   invitation.to_slack
                 rescue Slack::Web::Api::Errors::SlackError => e
                   case e.message
-                  when 'already_invited'
-                    { text: "User #{invitation.name_and_email} has already been invited." }
-                  when 'already_in_team_invited_user', 'already_in_team'
-                    { text: "User #{invitation.name_and_email} is already a member of the team." }
+                  when 'already_invited', 'already_in_team_invited_user'
+                    { text: "The user #{invitation.name_and_email} has already been invited to the team." }
+                  when 'already_in_team'
+                    { text: "The user #{invitation.name_and_email} is already a member of the team." }
                   else
                     { text: e.message }
                   end